BM/app/src/main/java/com/example/bm/transection.java
package com.example.bm;

public class transection {
    private String id;
    private String description;
    private String price;
    private String category;


    public transection() {
    }

    public transection(String id, String description, String price, String category) {
        this.id = id;
        this.description = description;
        this.price = price;
        this.category = category;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getPrice() {
        return price;
    }

    public void setPrice(String price) {
        this.price = price;
    }

    public String getCategory() {
        return category;
    }

    public void setCategory(String category) {
        this.category = category;
    }


}

